requirement.txt
websockets >= 10.1
pyyaml >= 6.0